requirements.txt
torch==2.0.0
numpy~=1.24.1
pillow~=9.3.0
matplotlib~=3.7.1
torchvision~=0.15.1+cu117
opencv-python~=4.7.0.72